created best offer component
